requirements.txt
ipython==7.12.0
pandas==1.0.1
geopandas==0.8.1
yatl==20200711.1
colour==0.1.5
pyyaml==5.3
plotly==4.8.2